feat: add more specific events to github and stackoverflow updates
