--- scrapper/src/main/java/edu/java/scrapper/client/sources/dto/GithubResponse.java
@@ -1,23 +1,0 @@
-package edu.java.scrapper.client.sources.dto;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.time.OffsetDateTime;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public record GithubResponse(
-    Long id,
-    String type,
-    Actor actor,
-    Repository repo,
-    OffsetDateTime createdAt
-) {
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public record Actor(
-        String displayLogin
-    ) {
-    }
-
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    public record Repository(String name) {
-    }
-}
